pyproject.toml
[project]
name = "py-api-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "luandnh",email = "luandnh98@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.11,<0.116.0)",
    "anyio (>=4.8.0,<5.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "sqlmodel (>=0.0.23,<0.0.24)",
    "aiomysql (>=0.2.0,<0.3.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
